% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annorm.R
\name{annorm}
\alias{annorm}
\title{Normalize CyTOF data panel}
\usage{
annorm(
  control_metadata_filename,
  control_data_filename,
  sample_metadata_filename,
  panel_filename,
  input_file_dir,
  val_file_dir = "none",
  output_file_dir,
  mode
)
}
\arguments{
\item{control_metadata_filename, }{control_data_filename, sample_metadata_filename, panel_filename, input_file_dir, val_file_dir,output_file_dir, mode}
}
\value{
normalized files with specified panel
}
\description{
This function normalize CyTOF data panel using external anchors.
}
