% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annorm_nrs.R
\name{annorm_nrs}
\alias{annorm_nrs}
\title{Normalize CyTOF data panel}
\usage{
annorm_nrs(
  sample_metadata_filename,
  panel_filename,
  input_file_dir,
  val_file_dir,
  output_file_dir,
  nchannels
)
}
\arguments{
\item{sample_metadata_filename, }{panel_filename, input_file_dir, val_file_dir,output_file_dir, nchannels}
}
\value{
normalized files with specified panel
}
\description{
This function normalize CyTOF data panel using internal anchors.
}
