% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anprep.R
\name{anprep}
\alias{anprep}
\title{Prep CyTOF control for normalization}
\usage{
anprep(metadata_filename, panel_filename, input_file_dir, output_file_dir)
}
\arguments{
\item{metadata_filename}{panel_filename input_file_dir output_file_dir}
}
\value{
data of summary statistics and concatenated control file
}
\description{
This function establish control reference for CyTOF data normalization.
}
