% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin-package.R
\docType{package}
\name{cytofin-package}
\alias{cytofin}
\alias{cytofin-package}
\title{cytofin: Integrate CyTOF Datasets From Heterogeneous Sources}
\description{
Integrate multiple CyTOF datasets collected from independent
  sources (i.e. labs, institutions, etc.). Cytofin performs CyTOF panel alignment
  across datasets ("homogenization") as well as batch correction using generalized
  anchors identified on each CyTOF plate ("normalization").
}
\author{
\strong{Maintainer}: Ben Lo \email{bennylo@stanford.edu}

Authors:
\itemize{
  \item Timothy Keyes \email{tkeyes@stanford.edu} (\href{https://orcid.org/0000-0003-0423-9679}{ORCID})
}

Other contributors:
\itemize{
  \item Kara Davis \email{kardavis@stanford.edu} [research team head, owner]
}

}
\keyword{internal}
