% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin.R
\name{cytofin}
\alias{cytofin}
\title{Title}
\usage{
cytofin(
  metadata_path,
  panel_path,
  input_data_path,
  output_data_path,
  verbose = FALSE
)
}
\arguments{
\item{metadata_path}{}

\item{panel_path}{}

\item{input_data_path}{}

\item{output_data_path}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
