% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_templates.R
\name{cytofin_generate_panel_template}
\alias{cytofin_generate_panel_template}
\title{Generate a template for a cytofin reference panel file}
\usage{
cytofin_generate_panel_template(
  file_name = "template_panel_info.csv",
  template_path = getwd()
)
}
\arguments{
\item{file_name}{A string representing the name of the .csv file to be
saved in the directory specified by \code{template_path}. Defaults to
"template_panel_info.csv"}

\item{template_path}{File path or connection where the template file should be
written. Defaults to the current working directory}
}
\description{
\code{cytofin_generate_panel_template} creates a template reference panel .csv file
(with the correct columns and dummy example data) in a specified location.
}
\examples{

# specify the path where you'd like to store the template file
my_name <- "panel_template.csv"
my_path <- file.path("~", "Desktop", "template_folder")

# generate the template file, which then can be edited manually 
cytofin_generate_panel_template(
   file_name = my_name, 
   template_path = my_path
)

}
