% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_homogenize.R
\name{cytofin_homogenize}
\alias{cytofin_homogenize}
\title{Homogenize CyTOF data panel}
\usage{
cytofin_homogenize(
  metadata_path,
  panel_path,
  input_data_path,
  output_data_path
)
}
\arguments{
\item{metadata_path}{A filepath leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names. Columns should include ____.
Must be in the current directory.}

\item{panel_path}{A filepath leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include ____.
Must be in the current directory.}

\item{input_data_path}{A folder directory containing the input CyTOF files
to be homogenized.}

\item{output_data_path}{A folder directory containing output homogenized files.}
}
\value{
\code{cytofin_homogenize} doesn't return anything. Instead, it has the
side-effect of saving homogenized files (in .fcs format) to the directory
specified with \code{output_data_path}. Each of the saved files will contain
homogenized, user-defined channels according to details specified in the
file at \code{panel_path.}
}
\description{
This function homogenizes CyTOF data (.fcs files) from heterogeneous sources
according to the standard panel in \code{panel_path.}
}
\examples{

# For a complete example of the `cytofin` workflow, 
# see the packages vignette by running `vignette(package = "cytofin")`

# download and store the metadata file
TO DO 

# download and store the panel file
TO DO 

# download and store the input data in a temporary directory
TO DO 

# specify a temporary output directory
TO DO 

# run homogenization
TO DO 
}
