% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_make_plots.R
\name{cytofin_make_plots}
\alias{cytofin_make_plots}
\title{Make diagnostic plots to evaluate CytofIn batch normalization}
\usage{
cytofin_make_plots(
  normalization_result,
  which_rows = 1:nrow(normalization_result),
  val_path = "none"
)
}
\arguments{
\item{normalization_result}{An output data.frame produced by the \code{cytofin_normalize} or
\code{cytofin_normalize_nrs} function.

The following columns should be present: \code{filename},
\code{cohort}, \code{plate_number}, \code{patient_id}, \code{condition}, \code{is_anchor}, \code{validation},
\code{universal_var}, \code{anchor_mean}, \code{anchor_var}, \code{mean_b4norm}, \code{var_b4norm},
\code{mean_norm}, \code{var_norm}, \code{mean_ctr_norm}, \code{var_ctr_norm}.}

\item{which_rows}{A numeric vector indicating which rows of \code{normalization_result}
(i.e. which .fcs files in the combined dataset) should be used for plotting. Defaults
to 1:nrow(normalization_result), which will make all possible plots.}

\item{val_path}{The folder directory containing validation (i.e. bead-normalized)
.fcs files corresponding to the input .fcs files in the metadata table. (Optional).}
}
\value{
8 diagnostic plots are made for each input .fcs file that was batch
normalized (i.e. each .fcs file represented as a row in \code{normalization_result}).
From left-to-right (and top-to-bottom), these plots represent the following:
\enumerate{
\item The entry in the universal mean vector corresponding to each antigen in the
consensus antigen panel. X-axis: antigen index in the universal mean vector.
Y-axis: Arcsinh-transformed entry in the universal mean vector corresponding
to each antigen.
\item The mean (across all cells) antigen expression vector for the anchor
associated with each input .fcs file both before and after normalization.
X-axis: antigen index (as in plot 1). Y-axis: Mean antigen expression in the
anchor .fcs file.
\item The mean (across all cells) antigen expression vector for each input
.fcs file both before and after normalization.
X-axis: antigen index (as in plot 1). Y-axis: Mean antigen expression in the
input .fcs file.
\item The mean (across all cells) antigen expression vector for each "validation"
(i.e. bead-normalized) .fcs file both before and after bead-normalization.
This plot can be used to compare CytofIn batch normalization with gold-
standard approaches. If \code{val_path} is "none", this plot will be identical to
plot 3 (see above).
X-axis: antigen index (as in plot 1). Y-axis: Mean antigen expression in the
validation .fcs file.
\item The entry in the universal standard deviation vector corresponding to each antigen in the
consensus antigen panel. X-axis: antigen index in the universal standard deviation vector.
Y-axis: Arcsinh-transformed entry in the universal standard deviation vector corresponding
to each antigen.
\item The standard deviation (across all cells) antigen expression vector for the anchor
associated with each input .fcs file both before and after normalization.
X-axis: antigen index (as in plot 1). Y-axis: the standard deviation of all
antigen expression values in the anchor .fcs file.
\item The standard deviation (across all cells) antigen expression vector for each input
.fcs file both before and after normalization.
X-axis: antigen index (as in plot 1). Y-axis: the standard deviation of all
antigen expression values in the input .fcs file.
\item The standard deviation (across all cells) antigen expression vector for each "validation"
(i.e. bead-normalized) .fcs file both before and after bead-normalization.
This plot can be used to compare CytofIn batch normalization with gold-
standard approaches. If \code{val_path} is "none", this plot will be identical to
plot 3 (see above).
X-axis: antigen index (as in plot 1). Y-axis: Standard deviation of all
antigen expression values in the validation .fcs file.
}
}
\description{
When given the output data structure from \code{cytofin_normalize} or \code{cytofin_normalize_nrs},
this function plots mean and variance plots for all normalized .fcs files and their
associated anchors.
}
