% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_normalize.R
\name{cytofin_normalize}
\alias{cytofin_normalize}
\title{Normalize CyTOF data panel}
\usage{
cytofin_normalize(
  control_metadata_path,
  anchor_statistics,
  sample_metadata_path,
  panel_path,
  input_path,
  val_path = "none",
  output_path,
  mode = c("meanshift", "meanshift_bulk", "variance", "z_score", "beadlike")
)
}
\arguments{
\item{control_metadata_path}{metadata table of anchor CyTOF files (.fcs)
(must be in the current directory).}

\item{anchor_statistics}{.rds object containing anchor reference statistics
(must be in the current directory).}

\item{sample_metadata_path}{metadata table of homogenized CyTOF files
(.fcs) (must be in the current directory).}

\item{panel_path}{standardized antigen panel table file (.xlsx/.csv)
(must be in the current directory).}

\item{input_path}{folder directory containing input homogenized CyTOF
data file.}

\item{val_path}{folder directory containing validation homogenized CyTOF
data file (optional).}

\item{output_path}{folder directory containing output normalized CyTOF data file.}

\item{mode}{transformation function used for normalization
(meanshift, meanshift_bulk, variance, z_score, or beadlike).}
}
\value{
normalized files with specified panel
}
\description{
This function normalize CyTOF data panel using external anchors.
}
