% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_normalize_nrs.R
\name{cytofin_normalize_nrs}
\alias{cytofin_normalize_nrs}
\title{Normalize CyTOF data panel}
\usage{
cytofin_normalize_nrs(
  sample_metadata_filename,
  panel_filename,
  input_file_dir,
  val_file_dir = "none",
  output_file_dir,
  nchannels
)
}
\arguments{
\item{sample_metadata_filename}{}

\item{panel_filename}{}

\item{input_file_dir}{}

\item{val_file_dir}{}

\item{output_file_dir}{}

\item{nchannels}{}
}
\value{
normalized files with specified panel
}
\description{
This function normalize CyTOF data panel using internal anchors.
}
