% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytofin_read_metadata}
\alias{cytofin_read_metadata}
\title{Read in a cytofin metadata file}
\usage{
cytofin_read_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{A filepath leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names. Columns should include
\code{filename}, \code{cohort}, \code{plate_number}, \code{patient_id}, \code{condition}, \code{population},
and \code{validation}. TO DO: Change the names of these columns to more descriptive
names and make sure that they are all actually needed.
See the vignette for details: \code{vignette("help", package = "cytofin")}}
}
\value{
A data.frame containing the metadata information in the
file stored at \code{metadata_path}.
}
\description{
This function reads a cytofin metadata file from a connection
that points to a .csv or a .xlsx file
}
\examples{
\dontrun{
my_path <- file.path("~", "foo", "bar", "metadata.csv")
my_metadata <- cytofin:::cytofin_read_metadata(my_path)
}
}
