% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytofin_read_panel_info}
\alias{cytofin_read_panel_info}
\title{Read in a cytofin reference panel information}
\usage{
cytofin_read_panel_info(panel_path)
}
\arguments{
\item{panel_path}{A file path leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include
\code{desc}, \code{range}, \code{metal_pattern}, \code{antigen_pattern}, \code{Lineage}, \code{Functional},
and \code{General}. TO DO: Change the names of these columns to more descriptive
names and make sure that they are all actually needed.
See the vignette for details: \code{vignette("help", package = "cytofin")}}
}
\value{
A data.frame containing the reference panel information in the
file stored at \code{panel_path}.
}
\description{
This function reads cytofin reference panel information from a connection
that points to a .csv or a .xlsx file
}
\examples{
my_path <- file.path("~", "foo", "bar", "panel.csv")
my_metadata <- cytofin_read_panel_info(my_path)

}
