% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_templates.R
\name{generate_metadata_template}
\alias{generate_metadata_template}
\title{Title}
\usage{
generate_metadata_template(template_path = getwd())
}
\arguments{
\item{template_path}{}
}
\value{

}
\description{
Title
}
