% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.R
\name{homogenize}
\alias{homogenize}
\title{Homogenize CyTOF data panel}
\usage{
homogenize(metadata_filename, panel_filename, input_file_dir, output_file_dir)
}
\arguments{
\item{metadata_filename}{panel_filename input_file_dir output_file_dir}
}
\value{
homogenized files with specified panel
}
\description{
This function homogenize CyTOF data panel given a standarize panel.
}
