% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{homogenize_flowFrame}
\alias{homogenize_flowFrame}
\title{Alter a flowFrame to only include data from channels in a reference panel}
\usage{
homogenize_flowFrame(fcs_raw, ref_panel)
}
\arguments{
\item{fcs_raw}{A flowFrame containing unprocessed CyTOF data}

\item{ref_panel}{A data.frame representing the reference panel data for a
cytofin analysis.}
}
\value{

}
\description{
Alter a flowFrame to only include data from channels in a reference panel
}
\examples{
TO DO: Insert examples
}
