% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_asinh}
\alias{rev_asinh}
\title{Reverses arcsinh transformation with cofactor \code{scale_factor} and a shift of \code{shift_factor}.}
\usage{
rev_asinh(x, shift_factor, scale_factor)
}
\arguments{
\item{x}{A numeric vector.}

\item{shift_factor}{The scalar value \code{a} in the following equation used to
transform CyTOF raw data ion counts using the hyperbolic arcsinh function:
\code{new_x <- asinh(a + b * x)}.}

\item{scale_factor}{The scalar value \code{b} in the following equation used to
transform CyTOF raw data ion counts using the hyperbolic arcsinh function:
\code{new_x <- asinh(a + b * x)}.}
}
\value{
A numeric vector after undergoing reverse
arcsinh transformation
}
\description{
Reverses arcsinh transformation with cofactor \code{scale_factor} and a shift of \code{shift_factor}.
}
