% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_homogenize.R
\name{cytofin_homogenize}
\alias{cytofin_homogenize}
\title{Homogenize CyTOF channels names using a consensus antigen panel}
\usage{
cytofin_homogenize(
  metadata_path,
  panel_path,
  input_data_path,
  output_data_path,
  prefix = "homogenized_",
  verbose = FALSE
)
}
\arguments{
\item{metadata_path}{A file path leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names in the first column (\code{filename})
and additional information about each .fcs file in subsequent columns.
Columns should include \code{filename}, \code{cohort}, \code{plate_number}, \code{patient_id},
\code{condition}, \code{is_anchor}, and \code{validation}.}

\item{panel_path}{A file path leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include
\code{metal_name}, \code{antigen_name}, \code{antigen_pattern},
\code{lineage}, \code{functional}, and \code{general}.}

\item{input_data_path}{A folder directory containing the input .fcs files
to be homogenized.}

\item{output_data_path}{A folder directory to which the output (i.e.
homogenized) .fcs files should be written.}

\item{prefix}{A string appended to the name of each input file to create the
name of the corresponding output file (post-homogenization). Defaults to
"homogenized_" (e.g. an input file named "file1.fcs" will correspond to
the output file "homogenized_file1.fcs" saved in \code{output_data_path}).}

\item{verbose}{A boolean value indicating whether progress message should be
printed to the console during homogenization. Defaults to FALSE.}
}
\value{
\code{cytofin_homogenize} doesn't return anything. Instead, it has the
side-effect of saving homogenized files (in .fcs format) to the directory
specified with \code{output_data_path}. Each of the saved files will contain
homogenized, user-defined channels according to details specified in the
file at \code{panel_path.}
}
\description{
This function homogenizes CyTOF data (.fcs files) from heterogeneous sources
according to the standard panel in a .csv file located at \code{panel_path.}
}
