% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_normalize_nrs.R
\name{cytofin_normalize_nrs}
\alias{cytofin_normalize_nrs}
\title{Normalize CyTOF data panel}
\usage{
cytofin_normalize_nrs(
  metadata_path,
  panel_path,
  input_data_path,
  output_data_path,
  input_prefix = "homogenized_",
  output_prefix = "normalized_",
  shift_factor = 0,
  scale_factor = 0.2,
  nchannels
)
}
\arguments{
\item{metadata_path}{A filepath leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names. Columns should include
\code{filename}, \code{cohort}, \code{plate_number}, \code{patient_id}, \code{condition}, \code{is_anchor},
and \code{validation}.
See the vignette for details: \code{vignette("help", package = "cytofin")}}

\item{panel_path}{A file path leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include
\code{metal_name}, \code{antigen_name}, \code{antigen_pattern},
\code{lineage}, \code{functional}, and \code{general}.
See the vignette for details: \code{vignette("help", package = "cytofin")}}

\item{input_data_path}{A folder directory containing the input CyTOF files
to be normalized. In most cases, this will be the directory to which the output
.fcs files from \code{cytofin_homogenize} were written.}

\item{output_data_path}{A folder directory to which the output (i.e.
batch normalized/batch corrected) .fcs files should be written.}

\item{input_prefix}{The string that was appended to the name of the input files
of \code{cytofin_homogenize} to create their corresponding output file names.
Defaults to "homogenized_".}

\item{output_prefix}{A string to be appended to the name of each input file
to create the name of the corresponding output file (post-homogenization).
Defaults to "normalized_" (e.g. an input file named "file1.fcs" will correspond to
the output file "normalized_file1.fcs" saved in \code{output_data_path}).}

\item{shift_factor}{The scalar value \code{a} in the following equation used to
transform CyTOF raw data ion counts using the hyperbolic arcsinh function:
\code{new_x <- asinh(a + b * x)}.
Defaults to 0.}

\item{scale_factor}{The scalar value \code{b} in the following equation used to
transform CyTOF raw data ion counts using the hyperbolic arcsinh function:
\code{new_x <- asinh(a + b * x)}.
Defaults to 0.2.}

\item{nchannels}{}

\item{val_file_dir}{}
}
\value{
Batch-normalized .fcs files are saved in the directory specified by
\code{output_data_path}.

In addition, a data.frame containing information about
each input .fcs file (that can be used for plotting with \code{cytofin_make_plots})
is returned with the following columns:
\itemize{
\item All of the columns in the input metadata table (located at \code{metadata_path})
\item \strong{universal_mean}: the universal mean vector to which all files are adjusted
(will be identical for all input .fcs files)
\item \strong{universal_var}: the universal mean vector to which all files are adjusted
(will be identical for all input .fcs files)
\item \strong{anchor_mean}: the mean (across all cells) vector for the anchor file associated
with each input .fcs file (i.e. the anchor located on the same plate as the
input .fcs file)
\item \strong{anchor_var}: the variance (across all cells) vector for the anchor file associated
with each input .fcs file (i.e. the anchor located on the same plate as the
input .fcs file)
\item \strong{mean_b4norm}: the mean (across all cells) vector of the input .fcs file
before batch normalization.
\item \strong{var_b4norm}: the variance (across all cells) vector of the input .fcs file
before batch normalization.
\item \strong{mean_norm}: the mean (across all cells) vector of the input .fcs file
after batch normalization.
\item \strong{var_norm}: the variance (across all cells) vector of the input .fcs file
after batch normalization.
}
}
\description{
This function normalize CyTOF data panel using internal anchors.
}
