% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofin_prep_anchors.R
\name{cytofin_prep_anchors}
\alias{cytofin_prep_anchors}
\title{Prepare CyTOF controls for batch normalization}
\usage{
cytofin_prep_anchors(metadata_path, panel_path, input_data_path)
}
\arguments{
\item{metadata_path}{A filepath leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names. Columns should include
\code{filename}, \code{cohort}, \code{plate_number}, \code{patient_id}, \code{condition}, \code{population},
and \code{validation}. TO DO: Change the names of these columns to more descriptive
names and make sure that they are all actually needed.
See the vignette for details: \code{vignette("help", package = "cytofin")}}

\item{panel_path}{A file path leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include
\code{desc}, \code{range}, \code{metal_pattern}, \code{antigen_pattern}, \code{Lineage}, \code{Functional},
and \code{General}. TO DO: Change the names of these columns to more descriptive
names and make sure that they are all actually needed.
See the vignette for details: \code{vignette("help", package = "cytofin")}}

\item{input_data_path}{A folder directory containing the input CyTOF files
to be prepped for normalization}
}
\value{
Data of summary statistics and concatenated control file
}
\description{
This function establishes control reference for CyTOF data normalization.
}
