% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.R
\name{homogenize}
\alias{homogenize}
\title{Homogenize CyTOF data panel}
\usage{
homogenize(metadata_filename, panel_filename, input_file_dir, output_file_dir)
}
\arguments{
\item{metadata_filename}{A filepath leading to an .xlsx or .csv file
containing a table of CyTOF file (.fcs file) names. Columns should include ____.
Must be in the current directory.}

\item{panel_filename}{A filepath leading to an .xlsx or .csv file containing
a table of standardized antigen panel information. Columns should include ____.
Must be in the current directory.}

\item{input_file_dir}{A folder directory containing the input CyTOF files
to be homogenized.}

\item{output_file_dir}{A folder directory containing output homogenized files.}
}
\value{
Homogenized files (in .fcs format) with user-defined channels according
to the file named panel_filename.
}
\description{
This function homogenizes CyTOF data (.fcs files) from heterogeneous sources
according to the standard panel in panel_filename.
}
